Reorganize JSON serializer implementation layout
