+++ src/compiler/serialization/impl/json/ucs-format-json.ts
@@ -1,0 +1,36 @@
+import { EsCode, EsSnippet, esline } from 'esgen';
+import { UcSchema } from '../../../../schema/uc-schema.js';
+import { UC_MODULE_SERIALIZER } from '../../../impl/uc-modules.js';
+import { UcsFormatter, UcsFormatterSignature } from '../../ucs-formatter.js';
+
+export function ucsFormatJSON<T>(formatter: UcsFormatter<T>): UcsFormatter<T> {
+  return (args, schema, context) => ucsCheckJSON(args, schema, formatter(args, schema, context));
+}
+
+export function ucsCheckJSON(
+  { writer, value }: UcsFormatterSignature.Values,
+  schema: UcSchema,
+  onValue: EsSnippet,
+  {
+    onNull = code => {
+      const jsonNull = UC_MODULE_SERIALIZER.import('UCS_JSON_NULL');
+
+      code.write(esline`await ${writer}.ready;`, esline`${writer}.write(${jsonNull})`);
+    },
+  }: {
+    readonly onNull?: EsSnippet;
+  } = {},
+): EsSnippet {
+  return function checkConstraints(code: EsCode) {
+    if (schema.nullable || schema.optional) {
+      code
+        .write(esline`if (${value} != null) {`)
+        .indent(onValue)
+        .write(`} else {`)
+        .indent(onNull)
+        .write('}');
+    } else {
+      code.write(onValue);
+    }
+  };
+}
